update many script.
