--- Base Defense (Zombie)/Assets/Scripts/EnemyMovement.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 36d83cd393f4e2e48b16cfa022743a1a
-timeCreated: 1434776719
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
