--- Base Defense (Zombie)/Assets/Scripts/GameLoader.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class GameLoader : MonoBehaviour {
-	public GameObject gameManager;			//GameManager prefab to instantiate.
-
-	void Awake ()
-	{
-		//Check if a GameManager has already been assigned to static variable GameManager.instance or if it's still null
-		if (GameManager.instance == null)
-
-			//Instantiate gameManager prefab
-			Instantiate(gameManager);
-	}
-}
