--- Base Defense (Zombie)/Assets/Scripts/GameLoader.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 229db7cfcde8bd54ea7a9feb35c00cba
-timeCreated: 1435013397
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
