--- Base Defense (Zombie)/Assets/Scripts/GameManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 861a593a9cf191144b8035448b9d8e48
-timeCreated: 1435012127
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
