--- Base Defense (Zombie)/Assets/Scripts/GreengruntHP.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8142824540574e645b7d9ac25cf27083
-timeCreated: 1435398337
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
