--- Base Defense (Zombie)/Assets/Scripts/HeroVision.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5b630bad597fb9f48832979d5634d339
-timeCreated: 1435039918
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
