--- Base Defense (Zombie)/Assets/Scripts/LinePatternMovement.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 972470d0e5ee0ff4dab23b46eed57e46
-timeCreated: 1435732850
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
