--- Base Defense (Zombie)/Assets/_Scenes/main.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8cda1978ea94bef428a5739501be5098
-timeCreated: 1434776183
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
